backend/requirements.txt
fastapi==0.105.0
uvicorn==0.24.0
pydantic==1.10.8
PyGithub==1.59.1
python-dotenv==1.0.0
requests==2.31.0
nltk==3.8.1
pytest==7.4.3
pylint==3.0.2
black==23.11.0
pyyaml==6.0.1
python-multipart==0.0.6
psutil==5.9.5
PyJWT==2.8.0
prometheus-client==0.17.1
fastapi-limiter==0.1.5 